backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
sqlalchemy>=2.0,<3
pydantic>=2.8,<3
pydantic-settings>=2.4,<3
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
email-validator>=2.2,<3
openai-chatkit
langchain-core>=0.3
langchain-community>=0.3
langchain-ollama>=0.2
langchain-qdrant>=0.2
qdrant-client>=1.12
fastembed>=0.4
